full-stack mvc done

+++ src/main/java/edu/uncg/csc340/animal_api/Panda.java
@@ -12,8 +12,9 @@ public class Panda {
     private String type;
 
 
-    public int getId() {
+    public long getId() {
         return id;
+
     }
     public void setId(int id) {
         this.id = id;
